requirements.txt
asgiref==3.6.0
Django==3.2
django-filter==1.1.0
django-simple-history==3.3.0
django-utils-six==2.0
djangorestframework==3.14.0
djangorestframework-filters==0.11.1
Markdown==3.4.3
Pillow==9.5.0
pytz==2023.3
sqlparse==0.4.4
